+++ CHANGELOG.rst
@@ -15,6 +15,10 @@ v34.9.5 (2025-02-19)
   with other aboutcode submodules.
   https://github.com/aboutcode-org/scancode.io/issues/1423
 
+- Add proper progress logging for the ``assemble`` section of the
+  ``scan_for_application_packages``.
+  https://github.com/aboutcode-org/scancode.io/issues/1601
+
 v34.9.4 (2025-01-21)
 --------------------